fix bug

+++ src/database/migrations/2019_10_16_100000_create_calendars_table.php
@@ -20,8 +20,10 @@ class CreateCalendarsTable extends Migration
             $table->timestamps();
         });
         Schema::table('events', function (Blueprint $table) {
-            $table->foreign('calendar_id')->references('id')->on('calendars')
-                ->onDelete('cascade');
+            if (Schema::hasColumn('events', 'calendar_id')) {
+                $table->foreign('calendar_id')->references('id')->on('calendars')
+                    ->onDelete('cascade');
+            }
         });
     }
